Generate device ID from discovered name when private MAC is used
New OS versions added the ability to randomize MAC addresses. With
iOS 14, this feature is now enabled by default, resulting in duplicated
device names in the logs.

To workaround this issue, when such MAC is detected, the discovered name
is used to generate the device ID instead of the dynamic MAC address.